services/api-gateway/requirements.txt
Flask==2.0.3
requests==2.26.0
PyJWT==2.3.0
Werkzeug==2.0.3
